Extract controls holder, move waveform display to sample player
