+++ src/components/audiofiles/ControlsHolder.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="controls" title="">
+    <slot />
+  </div>
+</template>
+
+<style scoped>
+  :is(.controls) {
+    display: flex;
+    padding: 8px;
+    align-items: center;
+    background: #121212;
+  }
+</style>
