+++ src/components/audiofiles/SamplePlayer.vue
@@ -3,6 +3,8 @@
   import type { AudioFile } from "@/stores/slices"
   import { useSlices } from "@/stores/slices"
   import ButtonControl from "@/components/audiofiles/ButtonControl.vue"
+  import ControlsHolder from "@/components/audiofiles/ControlsHolder.vue"
+  import SampleWaveform from "@/components/audiofiles/SampleWaveform.vue"
 
   const slicesStore = useSlices()
   const ctx = slicesStore.audioContext
@@ -26,6 +28,10 @@
     isPlaying.value = false
   }
 
+  function togglePlayback() {
+    isPlaying.value ? stop() : play()
+  }
+
   defineExpose({
     play,
     stop,
@@ -34,16 +40,34 @@
 </script>
 
 <template>
-  <ButtonControl
-    v-if="!isPlaying"
-    :title="`Play ${file.name}`"
-    icon="play_arrow"
-    @click="play"
-  />
-  <ButtonControl
-    v-if="isPlaying"
-    :title="`Stop playing ${file.name}`"
-    icon="stop"
-    @click="stop"
-  />
+  <SampleWaveform :file="file" @click="togglePlayback" />
+  <ControlsHolder class="controls">
+    <ButtonControl
+      v-if="!isPlaying"
+      :title="`Play ${file.name}`"
+      icon="play_arrow"
+      @click="play"
+    />
+    <ButtonControl
+      v-if="isPlaying"
+      :title="`Stop playing ${file.name}`"
+      icon="stop"
+      @click="stop"
+    />
+    <slot name="controls" />
+  </ControlsHolder>
 </template>
+
+<style scoped>
+  :is(.controls button) {
+    margin: 0 8px;
+  }
+
+  :is(.controls :first-child) {
+    margin-left: 0;
+  }
+
+  :is(.controls :last-child) {
+    margin-right: 0;
+  }
+</style>
